# feat(sqlite): Improve throughput of `SqliteEventCacheStore::load_all_chunks_metadata` by 1140%